Modify the policy logger start logic to fit in the ready metric.
This change is part of implementing the readiness SLI as described
in go/gke-policy-logging-metrics.

The policy logging feature is considered to be ready if both the
controller cache is synced and the policy logger pipeline are
setup correctly. The change sets up the policy logger pipeline at
the node start time, instead of being controlled by the NetworkLogging
CR. Policy events will be dropped early if logging is disabled.
This is aligned with our long-term vision that counter should be
supported regardless if logging is enabled or not.

Bug: b/166812557

Change-Id: I66bc7a1425b250e5916e0bd9b511f145f80227b1