Reacts to pod deletion locally for nodelocaldns
Current implementation of LRP relies on k8s control plane to react to
backend pod events. In large clusters, this causes a delay when
restoring DNS service to kube-dns upon the event of node-local-dns pods
(e.g., a rolling upgrade of node-local-dns) and this delay will lead to
DNS queires lost because the agent will still direct DNS packets to the
local backends which no longer exists and will only stop doing so when
it got the local backend deletion event from the api-server.

This PR invokes the service restoration logic in the local pod deletion
path. When kubelet issues cmdDel to Cilium-CNI, we still proactively
hint redirect service manager and let it decide whether we need to
restore original svc or not.

Manually tested in a small cluster.

BUG: 189861453

Change-Id: I559c3b062e368e731185d30221531f68f08ddaf8