[backport] Always insert CEPs into the largest available CES
Currently, with FCFS (First Come, First Served) batching mode, CiliumEndpoints are inserted into the first available (non-full, non-empty) CiliumEndpointSlice, which comes in an unordered list from the cache.

The change is to go through the list and pick the largest available
CES, instead of the first available. With this change, it’s guaranteed
that the creation of CEPs will cause the minimum number of CES
updates, using the FCFS batching mode.

This was backported from master to the v1.12 branch

Change-Id: Ie777fcfa6b832dea5f907c078b17f3e6a962e270
Signed-off-by: Dorde Lapcevic <dordel@google.com>
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/532726
Reviewed-by: Weilong Cui <cuiwl@google.com>
Unit-Verified: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Reviewed-by: Alan Kutniewski <kutniewski@google.com>