[backport from master] bpf,test: Add an additional optional pktgen stage to BPF tests
This commit adds the ability to define a `PKGEN(<progtype>, <name>)`
program along with the existing `SETUP` and `TEST` programs. This PKTGEN
program will be ran before the `SETUP` and/or `TEST`. The resulting
context is piped directly into the next programs.

The primary motivation of this addition is that the kernel will take
the context when `BPF_PROG_{test}_RUN` is invoked and will set some
fields in `__sk_buff` which would be 0 if we pass in an empty buffer.

Change-Id: Ie0afd495cf9f95475563be601f5b27bda8c20023
Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>
[jwi: add Docs update, and polish the hairpin test changes a bit]
Signed-off-by: Julian Wiedmann <jwi@isovalent.com>
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/613346
Reviewed-by: Maciej Skrocki <maciejskrocki@google.com>
Unit-Verified: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>