Add flag bpf-lb-preserve-world-id (defaults to true) to preserve backend compatibility in GDCH
In tunnel mode, the idenity of the external client is
set to world identity instead of the remote-node when
the traffic is sent to a node port and it bounces off
another node in the cluster. GDCH still relies on the
old behaviour where the client identity is set to
remote-node for this traffic so we are adding a flag.

Change-Id: I35620571e5c97dbed970a23a5ab4d5a766aade8f
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/590713
Unit-Verified: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Reviewed-by: Jonathan Hart <jonohart@google.com>