Log warning and allow static IP endpoints restoration
Multinic endpoints restoration expects the network information to be
present in the IPAM annotation, that is updated by ipam-controller via
cluster cidr configs. But in case of static IP endpoints, CCCs may not
exist and we do not support ipam mode in network spec yet.

We are allowing endpoints restoration for which the IPAM allocators are
absent and logging a warning mentioning the same.

BUG=b/263290429

Change-Id: Ibecf5e09fc9f6963d5fb5925128c218fcb37178b
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/597130
Unit-Verified: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Reviewed-by: Yifeng Shen <yfshen@google.com>
Reviewed-by: Bolun Zhao <blzhao@google.com>