service: Improve memory usage when handling update of a big service.
Change slices of Backends into slices of pointers to Backends.
That way the slice can be garbage collected even if some parts of
Backends are still in use.

Before this change the upsertService function had slice of backends.
This slice was assgined to service (svc.backends) and then the
notifyMonitorServiceUpsert function was used to push the even about the
upsert. The event created contained some fields from the backends.
Because of that the whole slice was kept in the memory.

After this fix the slice contains the pointers and they are not accessed
later so the slice can be removed from the memory. Only the accessed
fields need to be kept instead of whole objects.

Change-Id: Ibbd2362c56eb8c9e06e22f3936f3674ff87f32ea
Signed-off-by: Alan Kutniewski <kutniewski@google.com>
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/625510
Unit-Verified: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Reviewed-by: Dorde Lapcevic <dordel@google.com>