fix: Support clusters consisting of multiple node pools
It was erroneously assumed that there is only one node pool defining
contents of a cluster what caused constant stream of changes (nodes were
added and removed all the time).

In this change instead of storing a single map of all nodes, a map of
maps is used to preserve information on the cluster and node pools that
make it. All these maps are processed in an alphabetic order to
construct an update to node map. Differences between old and new mpa
are then sent to Hubble-relay as before.

Information from peer service (from the local cluster) are passed as is
and not reconciled with NodePools in any way. A map of local nodes is
used only to notify the late subscriber.

I also correct another error of mixing up the order of arguments to
cmp.Diff (it should be - expected, actual) with:
:%s/cmp.Diff\(([\w\.]*), ([\w\.]*)\)/cmp.Diff(\2, \1)/

Bug: b/283761697
Change-Id: If19c70513c6eb4b448e53858832b07bf116a2db0
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/774623
Reviewed-by: Nathan Perkins <nathanperkins@google.com>
Reviewed-by: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Unit-Verified: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Reviewed-by: Akhil Velagapudi <avelagap@google.com>